Question total (#4)
* Add persitent nfs volumne and remove firebase. Also add a guard to ensure that only authorized pages are viewed.

* Fix localStorage issue for logged guard.

* Fix cadriciel Dockerfile.

* changed deployments name.

* Add canValidate to userService

* Add docker nfs image.

* refactor structure.

* Has updated deployment files.

* fix error as we were using shutil instead of storage.

* Has updated deployment files.

* Fix an error for moodle file that was deleted while not existing.

* Provide moodle path to process copy.

* Change moodle path from places.

* has updated executor order.

* Fix a missing :

* Add some info in executor.

* uniformize requirements.txt packages' version.

* Authorize to delete any job at any time.

* add remove_tree to server Storage class.

* remove all files when deleting a job.

* Fix issue with executor: replace job by job_id. Add token to verify the id.

* Add admin_ prefix to all admin functions.

* Ensure UserService always use in a static mode in server.

* Delete also templates when deleting a user.

* Adding token for /job

* Catch all errors and print them for job_executor.

* Add all zip file when using moodle.

* Add an enum for checking role.

* Check role against enum.

* Change variables names when creating a new user.

* Fix Role.available().

* Fix admin commands.

* Fix delete tokens.

* Inherit admin features on signup.

* Add a default nginx config. Manage a group column in the csv file.

* Run process copy within process.

* Remove exception when only one box has been found.

* Remove all output zip files associated to a job.

* Fix executor Dockerfile.

* Change validation css.

* adapt style webapp task verification.

* Fix issue with pdf reader.

* Fixed verification style.

* Host fonts locally.

* Add some check in storage for server service.

* Ass also a check when moving a file with storage.

* Recognize slash in number and cut.

* Enforce right decimals for the grades.

* Enforce right decimals for the grades when not truncating. Add a bonus for truncated numbers.

* Sort numbers by probability.

* Remove mismatch.

+++ services/server/utils/storage.py
@@ -27,11 +27,15 @@ class Storage:
 
     def move_to(self, l_file, s_file):
         s_abs_file = self.abs_path(s_file)
+        if not os.path.exists(l_file):
+            raise ValueError("Storage can't find local file "+l_file)
         create_tree(s_abs_file)
         shutil.move(l_file, s_abs_file)
 
     def copy_from(self, s_file, l_file):
         s_abs_file = self.abs_path(s_file)
+        if not os.path.exists(s_abs_file):
+            raise ValueError("Storage can't find file "+s_abs_file)
         create_tree(l_file)
         shutil.copy(s_abs_file, l_file)
